Fix for Avalon/Avesta particle not attaching when used under Sacrifice.
I remember myself fixing this back when dun was dev, not sure why it got undone maybe it has associated bugs but I couldn't find any.

Fix is done on these 2 skills because the particles are critical to gameplay.

Anyway this fix can be applied to the rest of the game with some effort... albeit low returns.