Hecatic
Copied Hecatic into its own lua script so that it can make use of cast
filter. This should allow the game to check for the cast location before
attempting to cast. As well as realm checking so that it's not possible
for Medea to cast Hecatic at the last second to fly back into the
marble.